Asset Portfolio Template

A template for an asset portfolio used in ghgtools. You must follow the template structure in order for ghgtools to function properly.

Usage

AssetPortfolio

Format

AssetPortfolio

A data frame with 12 rows and 12 columns:

asset_id

Required input. Must be alpha-numeric. The asset_id is the primary key with activity data.

address

Optional input. Address of the asset.

city

Optional input. City of the asset.

state

Optional input. State of the asset.

zip

Required input for USA. Must be a numberic 5-digit zip code. Zip code is required for selecting the correct emisison factor for electricity in the USA.

country

Required input. Country of the asset.

sqft

Optional input. Square feet for buildings. Must be numeric value.

region

Optional input. Can be any value.

buisness_unit

Optional input. Can be any value.

asset_type

Required input. Must match an asset type in the Ecat_lookup table.

asset_subtype

Required input. Must match an asset subtype in the Ecat_lookup table.

year_built

Optional input. The year the asset was built.

...